feat: provider 추가

+++ src/main/java/com/smartspending/user/service/UserServiceImpl.java
@@ -1,12 +1,13 @@
 package com.smartspending.user.service;
 
+import com.smartspending.common.auth.jwt.JwtTokenProvider;
 import com.smartspending.common.exception.CommonResponseCode;
 import com.smartspending.common.exception.CustomException;
-import com.smartspending.common.auth.jwt.JwtTokenProvider;
 import com.smartspending.common.redis.RedisService;
 import com.smartspending.user.dto.request.*;
 import com.smartspending.user.dto.response.LoginResponseDto;
 import com.smartspending.user.entity.User;
+import com.smartspending.user.enums.Provider;
 import com.smartspending.user.repository.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.constraints.Email;
@@ -38,6 +39,11 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void sendVerificationCode(String email) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new CustomException(CommonResponseCode.USER_NOT_FOUND));
+        if (!user.getProvider().equals(Provider.LOCAL)) {
+            throw new CustomException(CommonResponseCode.USER_NOT_FOUND);
+        }
         String code = mailService.verifyCode();
         redisService.removeVerificationCode(email); // 이전 실패 경험 데이터 삭제
         redisService.saveVerificationCode(email, code);
@@ -76,6 +82,9 @@ public class UserServiceImpl implements UserService {
         User user = userRepository.findByEmail(requestDto.getEmail())
                 .orElseThrow(() -> new CustomException(CommonResponseCode.USER_NOT_FOUND));
 
+        if (!user.getProvider().equals(Provider.LOCAL)) {
+            throw new CustomException(CommonResponseCode.USER_NOT_FOUND);
+        }
         if (!passwordEncoder.matches(requestDto.getPassword(), user.getPassword())) {
             throw new CustomException(CommonResponseCode.USER_NOT_FOUND);
         }
